Refactor agent-canvas-renderer: split 2,812 lines into 5 specialized modules
- Extract renderer-primitives.js (179 lines) - basic drawing utilities
- Extract renderer-connections.js (387 lines) - all connection rendering
- Extract renderer-nodes.js (719 lines) - task, agent, combiner nodes
- Extract renderer-panels.js (938 lines) - detail panels and timeline
- Extract renderer-ui.js (672 lines) - buttons, notifications, menus
- Main renderer now orchestrator (167 lines, 94% reduction)
- All modules under 1000 lines ✓
